user-service/requirements.txt
fastapi>=0.109.0  # Version minimale compatible avec Pydantic v2
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
redis==5.0.1
prometheus-client==0.19.0
httpx==0.25.2
pytest==7.4.0